+++ src/components/Header/Header.js
@@ -7,6 +7,7 @@ const authenticatedOptions = (
     <Nav.Link href="#change-password">Change Password</Nav.Link>
     <Nav.Link href="#sign-out">Sign Out</Nav.Link>
     <Nav.Link href="#products">Products</Nav.Link>
+    <Nav.Link href="#account">Account</Nav.Link>
   </Fragment>
 )
